requirements.txt
asgiref==3.6.0
boto3==1.26.110
botocore==1.29.110
Django==4.2
django-cors-headers==3.14.0
django-restframework==0.0.1
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
